fix smoke tests
